# pkg/api/v1/docs/api_swagger.yaml
basePath: /v1
consumes:
- application/json
definitions:
...
  models.ClientCertificateResponse:
    properties:
      additional_dns_names:
        items:
          type: string
        type: array
      additional_ip_addresses:
        items:
          type: string
        type: array
      authority_name:
        type: string
      ca_certificate:
        format: base64
        type: string
      expiry:
        type: string
      public_certificate:
        format: base64
        type: string
      template_name:
        type: string
    type: object
...
  models.CreateClientCertificateRequest:
    properties:
      additional_dns_names:
        items:
          type: string
        type: array
      additional_ip_addresses:
        items:
          type: string
        type: array
      csr:
        format: base64
        type: string
      template_name:
        type: string
    type: object
...
  models.CreateServerCertificateRequest:
    properties:
      additional_dns_names:
        items:
          type: string
        type: array
      additional_ip_addresses:
        items:
          type: string
        type: array
      csr:
        format: base64
        type: string
      template_name:
        type: string
    type: object
...
  models.ServerCertificateResponse:
    properties:
      additional_dns_names:
        items:
          type: string
        type: array
      additional_ip_addresses:
        items:
          type: string
        type: array
      authority_name:
        type: string
      ca_certificate:
        format: base64
        type: string
      expiry:
        type: string
      public_certificate:
        format: base64
        type: string
      template_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: admin@loopholelabs.io
    name: API Support
  description: Returns the health and status of the various services that make up
    the API.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://loopholelabs.io/privacy
  title: EndKey API V1
  version: "1.0"
paths:
...
  /certificate/client:
    post:
      consumes:
      - application/json
      description: Create a new Client Certificate
      parameters:
      - description: Create Client Certificate Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateClientCertificateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ClientCertificateResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "412":
          description: Precondition Failed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      tags:
      - certificate
  /certificate/server:
    post:
      consumes:
      - application/json
      description: Create a new Server Certificate
      parameters:
      - description: Create Server Certificate Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateServerCertificateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ServerCertificateResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "412":
          description: Precondition Failed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      tags:
      - certificate
...
produces:
- application/json
schemes:
- https
swagger: "2.0"